refactor: NavigateType 패키지 이동
