--- app/src/main/java/woowacourse/movie/domain/model/NavigateType.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.domain.model
-
-import java.io.Serializable
-
-enum class NavigateType : Serializable {
-    NONE,
-    NAVIGATE_TO_MAIN,
-    NAVIGATE_TO_PREVIOUS,
-}
