+++ app/src/main/java/woowacourse/movie/feature/model/NavigateType.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.feature.model
+
+enum class NavigateType {
+    NONE,
+    NAVIGATE_TO_MAIN,
+    NAVIGATE_TO_PREVIOUS,
+}
